+++ LearningDirectX2/Entity.h
@@ -80,7 +80,7 @@ public:
 	virtual void Update(double dt);
 	virtual void Render();
 
-	virtual void OnCollision(Entity *impactor, SideCollision side, float collisionTime, float dt = (1.0/60));
+	virtual void OnCollision(Entity *impactor, SideCollision side, float collisionTime, double dt = (1.0/60));
 	virtual EarnedData OnDestroy();
 
 protected:
