+++ LearningDirectX2/Flames.cpp
@@ -40,7 +40,7 @@ void Flames::Update(double dt) {
 	Weapon::Update(dt);
 }
 
-void Flames::OnCollision(Entity * impactor, Entity::SideCollision side, float collisionTime, float dt) {
+void Flames::OnCollision(Entity * impactor, Entity::SideCollision side, float collisionTime, double dt) {
 	auto impactorType = impactor->GetType();
 	if (impactorType == Layer::EnemyType || impactorType == Layer::ItemHolderType || impactorType == Layer::EProjectileType) {
 		impactor->OnDestroy();
