+++ LearningDirectX2/Flames.h
@@ -9,7 +9,7 @@ public:
 	Flames();
 	~Flames();
 	virtual void Update(double dt);
-	virtual void OnCollision(Entity *impactor, Entity::SideCollision side, float collisionTime, float dt = 1.0/60);
+	virtual void OnCollision(Entity *impactor, Entity::SideCollision side, float collisionTime, double dt = 1.0/60);
 	virtual void Instantiate(D3DXVECTOR3 position);
 	virtual EarnedData OnDestroy();
 private:
