+++ LearningDirectX2/Knife.cpp
@@ -22,7 +22,7 @@ Knife::Knife() : Weapon() {
 	point = 100;
 }
 
-void Knife::OnCollision(Entity * impactor, Entity::SideCollision side, float collisionTime, float dt) {
+void Knife::OnCollision(Entity * impactor, Entity::SideCollision side, float collisionTime, double dt) {
 }
 
 void Knife::Update(double dt) {
