+++ LearningDirectX2/Soldier.cpp
@@ -22,7 +22,7 @@ Soldier::~Soldier() {
 	soldierAttackState = NULL;
 }
 
-void Soldier::OnCollision(Entity * impactor, Entity::SideCollision side, float collisionTime, float dt) {
+void Soldier::OnCollision(Entity * impactor, Entity::SideCollision side, float collisionTime, double dt) {
 	Enemy::OnCollision(impactor, side, collisionTime);
 	if (impactor->GetType() == Layer::StaticType && side == Bottom)
 		onGround = true;
